Squashed commit of the following:
Make sure the cache directory is local to the project.
This is needed to support workspace styles features in editors like VSCode, for instance you may have a project with an elixir backend in /myproject/backend, and Elm code in /myproject/frontend/, in the editor you might open /myproject/ to jump between projects easier.
This adds a new LocalCache class that takes the project path as an argument and stores the elm-analyse cache relative to that.